refactor: changes filename to utils
